+++ R/complete.R
@@ -57,3 +57,8 @@ complete_.data.frame <- function(data, cols, fill = list(), ...) {
 
   full
 }
+
+#' @export
+complete_.grouped_df <- function(data, cols, fill = list(), ...) {
+  dplyr::grouped_df(NextMethod(), dplyr::groups(data))
+}